created Auth component and linked sign-up route to it. Also moved posts and new-post jsx files to pages dir. #3
