+++ client/pages/new-post.jsx
