+++ client/pages/posts.jsx
